+++ src/Maintenance/Maintenance.Test/ScenarioTests/MaintenanceTestRunner.cs
@@ -37,7 +37,8 @@ namespace Microsoft.Azure.Commands.Maintenance.Test.ScenarioTests
                     helper.RMProfileModule,
                     helper.GetRMModulePath("Az.Maintenance.psd1"),
                     helper.GetRMModulePath("Az.Compute.psd1"),
-                    helper.GetRMModulePath("Az.Network.psd1")
+                    helper.GetRMModulePath("Az.Network.psd1"),
+                    helper.GetRMModulePath("Az.ResourceGraph.psd1")
                 })
                 .WithNewRecordMatcherArguments(
                     userAgentsToIgnore: new Dictionary<string, string>
@@ -51,7 +52,8 @@ namespace Microsoft.Azure.Commands.Maintenance.Test.ScenarioTests
                         { "Microsoft.Features", null },
                         { "Microsoft.Authorization", null },
                         { "Microsoft.Maintenance", null },
-                        { "Microsoft.Network", null }
+                        { "Microsoft.Network", null },
+                        { "Microsoft.ResourceGraph", null }
                     }
                 )
                 .Build();