+++ src/Maintenance/Maintenance/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Added support to acknowledge ScheduledEvents using 'Set-AzScheduledEvents' and 'Set-AzScheduledEventsList.'
 
 ## Version 1.6.0
 * Added ChangeSafety Support